Use embeddable Python for launcher
